@@ -16,8 +16,9 @@ namespace MinecraftService.Shared.Utilities
     public class MinecraftFileUtilities {
         public static bool UpdateWorldPackFile(string filePath, PackManifestJsonModel manifest) {
             WorldPackFileModel worldPackFile = new(filePath);
-            if (worldPackFile.Contents.Where(x => x.pack_id == manifest.header.uuid).Count() > 0) {
-                return false;
+            var foundpack = worldPackFile.Contents.Where(x => x.pack_id == manifest.header.uuid).ToList();
+            if (foundpack.Count() == 1) {
+                foundpack[0] = new WorldPackEntryJsonModel(manifest.header.uuid, manifest.header.version);
             }
             worldPackFile.Contents.Add(new WorldPackEntryJsonModel(manifest.header.uuid, manifest.header.version));
             worldPackFile.SaveFile();
@@ -37,20 +38,18 @@ namespace MinecraftService.Shared.Utilities
 
         public static Task ValidateFixWorldPackFiles(string serverPath, string levelName) {
             return Task.Run(() => {
-                MinecraftPackParser resourceParser = new();
-                MinecraftPackParser behaviorParser = new();
+                MinecraftPackParser packParser = new();
                 try {
-                    resourceParser.ParseDirectory(GetServerDirectory(ServerDirectoryKeys.ResourcePacksDir, serverPath, levelName), 0);
-                    behaviorParser.ParseDirectory(GetServerDirectory(ServerDirectoryKeys.BehaviorPacksDir, serverPath, levelName), 0);
+                    packParser.ParseDirectory(GetServerDirectory(ServerDirectoryKeys.ServerWorldDir_LevelName, serverPath, levelName), 0);
                     List<WorldPackEntryJsonModel> resourceJsonList = JsonConvert.DeserializeObject<List<WorldPackEntryJsonModel>>(File.ReadAllText(GetServerFilePath(ServerFileNameKeys.WorldResourcePacks, serverPath, levelName)));
                     List<WorldPackEntryJsonModel> behaviorJsonList = JsonConvert.DeserializeObject<List<WorldPackEntryJsonModel>>(File.ReadAllText(GetServerFilePath(ServerFileNameKeys.WorldBehaviorPacks, serverPath, levelName)));
                     foreach (WorldPackEntryJsonModel entry in resourceJsonList) {
-                        if (resourceParser.FoundPacks.Count(x => x.JsonManifest.header.uuid == entry.pack_id) == 0) {
+                        if (packParser.FoundPacks.Count(x => x.JsonManifest.header.uuid == entry.pack_id) == 0) {
                             RemoveEntryFromWorldPackFile(GetServerFilePath(ServerFileNameKeys.WorldResourcePacks, serverPath, levelName), entry).Wait();
                         }
                     }
                     foreach (WorldPackEntryJsonModel entry in behaviorJsonList) {
-                        if (resourceParser.FoundPacks.Count(x => x.JsonManifest.header.uuid == entry.pack_id) == 0) {
+                        if (packParser.FoundPacks.Count(x => x.JsonManifest.header.uuid == entry.pack_id) == 0) {
                             RemoveEntryFromWorldPackFile(GetServerFilePath(ServerFileNameKeys.WorldBehaviorPacks, serverPath, levelName), entry).Wait();
                         }
                     }
@@ -112,49 +111,43 @@ namespace MinecraftService.Shared.Utilities
             File.WriteAllLines(GetServerFilePath(ServerFileNameKeys.ServerProps, server), output);
         }
 
-        public static List<Property> GetDefaultPropListFromFile(string filePath) {
-            List<string[]> fileContents = FilterLinesFromPropFile(filePath);
-            List<Property> result = new List<Property>();
-            fileContents.ForEach(prop => {
-                if (prop.Length == 1) {
-                    prop = new string[] { prop[0], "" };
-                }
-                result.Add(new Property(prop[0], prop[1]));
-            });
-            return result;
-        }
-
         public static List<Property> CopyPropList(List<Property> souceList) {
-            List<Property> result = new List<Property>();
+            List<Property> result = [];
             foreach (Property prop in souceList) {
                 result.Add(new Property(prop));
             }
             return result;
         }
 
-        public static List<string[]> FilterLinesFromPlayerDbFile(string filePath) {
+        public static List<string[]> FilterLinesFromPropFile(string filePath, char delim = '=') {
             FileUtilities.CreateInexistantFile(filePath);
             return File.ReadAllLines(filePath)
                     .Where(x => !x.StartsWith("#"))
                     .Where(x => !string.IsNullOrWhiteSpace(x))
-                    .Select(x => x.Split(','))
+                    .Select(x => x.Split(delim))
                     .ToList();
         }
 
-        public static List<string[]> FilterLinesFromPropFile(string filePath) {
-            FileUtilities.CreateInexistantFile(filePath);
-            return File.ReadAllLines(filePath)
-                    .Where(x => !x.StartsWith("#"))
-                    .Where(x => !string.IsNullOrWhiteSpace(x))
-                    .Select(x => x.Split('='))
-                    .ToList();
+        public static List<Property> CreatePropListFromFile(string filePath, char delim = '=') {
+            List<string[]> result = FilterLinesFromPropFile(filePath, delim);
+            List<Property> output = [];
+            result.ForEach(x => {
+                if (x.Length < 2) {
+                    string[] replace = new string[2];
+                    replace[0] = x[0];
+                    x = replace;
+                }
+                output.Add(new(x[0], x[1]));
+            });
+            return output;
         }
+
 
         public static void CleanBedrockDirectory(IServerConfiguration server) {
             DirectoryInfo bedrockDir = new DirectoryInfo(server.GetSettingsProp(ServerPropertyKeys.ServerPath).ToString());
             Progress<ProgressModel> nullProgress = new();
-            FileUtilities.DeleteFilesFromDirectory($@"{server.GetSettingsProp(ServerPropertyKeys.ServerPath)}\resource_packs", true, nullProgress);
-            FileUtilities.DeleteFilesFromDirectory($@"{server.GetSettingsProp(ServerPropertyKeys.ServerPath)}\behavior_packs", true, nullProgress);
+            FileUtilities.DeleteFilesFromDirectory($@"{server.GetSettingsProp(ServerPropertyKeys.ServerPath)}\worlds\{server.GetProp(MmsDependServerPropKeys.LevelName)}\resource_packs", true, nullProgress);
+            FileUtilities.DeleteFilesFromDirectory($@"{server.GetSettingsProp(ServerPropertyKeys.ServerPath)}\worlds\{server.GetProp(MmsDependServerPropKeys.LevelName)}\behavior_packs", true, nullProgress);
             foreach (FileInfo file in bedrockDir.EnumerateFiles()) {
                 if (file.Extension.Equals(".exe") || file.Extension.Equals(".dll")) {
                     File.Delete(file.FullName);
