@@ -2,10 +2,8 @@
 using Newtonsoft.Json;
 using System;
 
-namespace MinecraftService.Shared.Classes.Server
-{
-    public class Player
-    {
+namespace MinecraftService.Shared.Classes.Server {
+    public class Player {
         [JsonProperty]
         private string Username { get; set; }
         [JsonProperty]
@@ -26,8 +24,7 @@ namespace MinecraftService.Shared.Classes.Server
         private bool IgnorePlayerLimits { get; set; }
 
         [JsonConstructor]
-        public Player(string playerID, string username, long firstConn, long lastConn, long lastDiscon, bool whtlist, string perm, bool ignoreLimit)
-        {
+        public Player(string playerID, string username, long firstConn, long lastConn, long lastDiscon, bool whtlist, string perm, bool ignoreLimit) {
             Username = username;
             PlayerID = playerID;
             FirstConnectedTime = firstConn;
@@ -38,20 +35,17 @@ namespace MinecraftService.Shared.Classes.Server
             IgnorePlayerLimits = ignoreLimit;
         }
 
-        public Player(string serverDefaultPermission)
-        {
+        public Player(string serverDefaultPermission) {
             ServerDefaultPerm = serverDefaultPermission;
             PermissionLevel = serverDefaultPermission;
         }
 
-        public Player()
-        {
+        public Player() {
             ServerDefaultPerm = "2";
             PermissionLevel = "2";
         }
 
-        public Player Initialize(string xuid, string username)
-        {
+        public Player Initialize(string xuid, string username) {
             PlayerID = xuid;
             Username = username;
             FirstConnectedTime = DateTime.Now.Ticks;
@@ -62,8 +56,7 @@ namespace MinecraftService.Shared.Classes.Server
 
         public string GetUsername() => Username;
 
-        public string SearchForProperty(string input)
-        {
+        public string SearchForProperty(string input) {
             input = input.ToLower();
             if (input == "name" || input == "username" || input == "un")
                 return Username.ToLower();
@@ -82,44 +75,35 @@ namespace MinecraftService.Shared.Classes.Server
 
         public string GetPlayerID() => PlayerID;
 
-        public (long First, long Conn, long Disconn) GetTimes()
-        {
+        public (long First, long Conn, long Disconn) GetTimes() {
             return (FirstConnectedTime, LastConnectedTime, LastDisconnectTime);
         }
 
-        public void UpdateTimes(long conn, long disconn)
-        {
+        public void UpdateTimes(long conn, long disconn) {
             LastConnectedTime = conn;
             LastDisconnectTime = disconn;
         }
 
-        public bool IsDefaultRegistration()
-        {
+        public bool IsDefaultRegistration() {
             return Whitelisted == false && IgnorePlayerLimits == false && PermissionLevel == ServerDefaultPerm;
         }
 
-        public string ToString(string format)
-        {
-            if (format == "Known")
-            {
+        public string ToString(string format) {
+            if (format == "Known") {
                 return $"{PlayerID},{Username},{FirstConnectedTime},{LastConnectedTime},{LastDisconnectTime}";
             }
-            if (format == "Registered")
-            {
+            if (format == "Registered") {
                 return $"{PlayerID},{Username},{PermissionLevel},{Whitelisted},{IgnorePlayerLimits}";
             }
             return null;
         }
 
-        public Player UpdatePlayerFromDbStrings(string[] dbString)
-        {
-            if (dbString == null || dbString[0] != PlayerID)
-            {
+        public Player UpdatePlayerFromDbStrings(string[] dbString) {
+            if (dbString == null || (PlayerID != null && dbString[0] != PlayerID)) {
                 throw new ArgumentException("Input null or Player update attempted with incorrect xuid!");
             }
             Username = dbString[1];
-            if (!long.TryParse(dbString[2], out long first) || !long.TryParse(dbString[3], out long conn) || !long.TryParse(dbString[4], out long disconn))
-            {
+            if (!long.TryParse(dbString[2], out long first) || !long.TryParse(dbString[3], out long conn) || !long.TryParse(dbString[4], out long disconn)) {
                 throw new InvalidOperationException("Could not parse player times, check logs!");
             }
             FirstConnectedTime = first;
@@ -128,16 +112,13 @@ namespace MinecraftService.Shared.Classes.Server
             return this;
         }
 
-        public Player UpdatePlayerFromRegStrings(string[] regString)
-        {
-            if (regString == null || regString[0] != PlayerID)
-            {
+        public Player UpdatePlayerFromRegStrings(string[] regString) {
+            if (regString == null || (PlayerID != null && regString[0] != PlayerID)) {
                 throw new ArgumentException("Input null or Player update attempted with incorrect xuid!");
             }
             Username = regString[1];
             PermissionLevel = regString[2];
-            if (!bool.TryParse(regString[3], out bool whiteList) || !bool.TryParse(regString[4], out bool ignoreLimits))
-            {
+            if (!bool.TryParse(regString[3], out bool whiteList) || !bool.TryParse(regString[4], out bool ignoreLimits)) {
                 throw new InvalidOperationException("Could not parse registration bools, check configs!");
             }
             Whitelisted = whiteList;
@@ -151,15 +132,13 @@ namespace MinecraftService.Shared.Classes.Server
 
         public string GetPermissionLevel() => PermissionLevel;
 
-        public override bool Equals(object obj)
-        {
+        public override bool Equals(object obj) {
             return obj is Player player &&
                    Username == player.Username &&
                    PlayerID == player.PlayerID;
         }
 
-        public override int GetHashCode()
-        {
+        public override int GetHashCode() {
             return HashCode.Combine(Username, PlayerID);
         }
     }
